Reorganize the tests with CUDA, Enzyme and Zygote
