+++ test/Project.toml
@@ -1,6 +1,4 @@
 [deps]
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 ManualNLPModels = "30dfa513-9b2f-4fb3-9796-781eabac1617"
@@ -11,11 +9,8 @@ ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatrixColorings = "0a514795-09f3-496d-8182-132a7b665d35"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-CUDA = "4, 5"
-Enzyme = "0.10, 0.11, 0.12"
 ForwardDiff = "0.10"
 ManualNLPModels = "0.1"
 NLPModels = "0.21"
@@ -23,4 +18,3 @@ NLPModelsModifiers = "0.7"
 NLPModelsTest = "0.10"
 ReverseDiff = "1"
 SparseMatrixColorings = "0.4.0"
-Zygote = "0.6"
