+++ test/runtests.jl
@@ -1,4 +1,4 @@
-using CUDA, LinearAlgebra, SparseArrays, Test
+using LinearAlgebra, SparseArrays, Test
 using SparseMatrixColorings
 using ADNLPModels, ManualNLPModels, NLPModels, NLPModelsModifiers, NLPModelsTest
 using ADNLPModels:
@@ -40,109 +40,7 @@ for problem in NLPModelsTest.nls_problems
   include("nls/problems/$(lowercase(problem)).jl")
 end
 
-# Additional backends used for tests
-push!(
-  ADNLPModels.predefined_backend,
-  :zygote_backend => Dict(
-    :gradient_backend => ADNLPModels.ZygoteADGradient,
-    :jprod_backend => ADNLPModels.ZygoteADJprod,
-    :jtprod_backend => ADNLPModels.ZygoteADJtprod,
-    :hprod_backend => ADNLPModels.ForwardDiffADHvprod,
-    :jacobian_backend => ADNLPModels.ZygoteADJacobian,
-    :hessian_backend => ADNLPModels.ZygoteADHessian,
-    :ghjvprod_backend => ADNLPModels.ForwardDiffADGHjvprod,
-    :jprod_residual_backend => ADNLPModels.ZygoteADJprod,
-    :jtprod_residual_backend => ADNLPModels.ZygoteADJtprod,
-    :hprod_residual_backend => ADNLPModels.ForwardDiffADHvprod,
-    :jacobian_residual_backend => ADNLPModels.ZygoteADJacobian,
-    :hessian_residual_backend => ADNLPModels.ZygoteADHessian,
-  ),
-)
-
-ReverseDiffAD(nvar, f) = ADNLPModels.ADModelBackend(
-  nvar,
-  f,
-  gradient_backend = ADNLPModels.ReverseDiffADGradient,
-  hprod_backend = ADNLPModels.ReverseDiffADHvprod,
-  jprod_backend = ADNLPModels.ReverseDiffADJprod,
-  jtprod_backend = ADNLPModels.ReverseDiffADJtprod,
-  jacobian_backend = ADNLPModels.ReverseDiffADJacobian,
-  hessian_backend = ADNLPModels.ReverseDiffADHessian,
-)
-
-function test_getter_setter(nlp)
-  @test get_adbackend(nlp) == nlp.adbackend
-  if typeof(nlp) <: ADNLPModel
-    set_adbackend!(nlp, ReverseDiffAD(nlp.meta.nvar, nlp.f))
-  elseif typeof(nlp) <: ADNLSModel
-    function F(x; nequ = nlp.nls_meta.nequ)
-      Fx = similar(x, nequ)
-      nlp.F!(Fx, x)
-      return Fx
-    end
-    set_adbackend!(nlp, ReverseDiffAD(nlp.meta.nvar, x -> sum(F(x) .^ 2)))
-  end
-  @test typeof(get_adbackend(nlp).gradient_backend) <: ADNLPModels.ReverseDiffADGradient
-  @test typeof(get_adbackend(nlp).hprod_backend) <: ADNLPModels.ReverseDiffADHvprod
-  @test typeof(get_adbackend(nlp).hessian_backend) <: ADNLPModels.ReverseDiffADHessian
-  set_adbackend!(
-    nlp,
-    gradient_backend = ADNLPModels.ForwardDiffADGradient,
-    jtprod_backend = ADNLPModels.GenericForwardDiffADJtprod(),
-  )
-  @test typeof(get_adbackend(nlp).gradient_backend) <: ADNLPModels.ForwardDiffADGradient
-  @test typeof(get_adbackend(nlp).hprod_backend) <: ADNLPModels.ReverseDiffADHvprod
-  @test typeof(get_adbackend(nlp).jtprod_backend) <: ADNLPModels.GenericForwardDiffADJtprod
-  @test typeof(get_adbackend(nlp).hessian_backend) <: ADNLPModels.ReverseDiffADHessian
-end
-
-ZygoteAD() = ADNLPModels.ADModelBackend(
-  ADNLPModels.ZygoteADGradient(),
-  ADNLPModels.GenericForwardDiffADHvprod(),
-  ADNLPModels.ZygoteADJprod(),
-  ADNLPModels.ZygoteADJtprod(),
-  ADNLPModels.ZygoteADJacobian(0),
-  ADNLPModels.ZygoteADHessian(0),
-  ADNLPModels.ForwardDiffADGHjvprod(),
-  ADNLPModels.EmptyADbackend(),
-  ADNLPModels.EmptyADbackend(),
-  ADNLPModels.EmptyADbackend(),
-  ADNLPModels.EmptyADbackend(),
-  ADNLPModels.EmptyADbackend(),
-)
-
-function test_autodiff_backend_error()
-  @testset "Error without loading package - $backend" for backend in [:ZygoteAD]
-    adbackend = eval(backend)()
-    @test_throws ArgumentError gradient(adbackend.gradient_backend, sum, [1.0])
-    @test_throws ArgumentError gradient!(adbackend.gradient_backend, [1.0], sum, [1.0])
-    @test_throws ArgumentError jacobian(adbackend.jacobian_backend, identity, [1.0])
-    @test_throws ArgumentError hessian(adbackend.hessian_backend, sum, [1.0])
-    @test_throws ArgumentError Jprod!(
-      adbackend.jprod_backend,
-      [1.0],
-      [1.0],
-      identity,
-      [1.0],
-      Val(:c),
-    )
-    @test_throws ArgumentError Jtprod!(
-      adbackend.jtprod_backend,
-      [1.0],
-      [1.0],
-      identity,
-      [1.0],
-      Val(:c),
-    )
-  end
-end
-
-# Test the argument error without loading the packages
-test_autodiff_backend_error()
-
-# Automatically loads the code for Zygote with Requires
-import Zygote
-
+include("utils.jl")
 include("nlp/basic.jl")
 include("nls/basic.jl")
 include("nlp/nlpmodelstest.jl")
