+++ test/utils.jl
@@ -1,0 +1,36 @@
+ReverseDiffAD(nvar, f) = ADNLPModels.ADModelBackend(
+  nvar,
+  f,
+  gradient_backend = ADNLPModels.ReverseDiffADGradient,
+  hprod_backend = ADNLPModels.ReverseDiffADHvprod,
+  jprod_backend = ADNLPModels.ReverseDiffADJprod,
+  jtprod_backend = ADNLPModels.ReverseDiffADJtprod,
+  jacobian_backend = ADNLPModels.ReverseDiffADJacobian,
+  hessian_backend = ADNLPModels.ReverseDiffADHessian,
+)
+
+function test_getter_setter(nlp)
+  @test get_adbackend(nlp) == nlp.adbackend
+  if typeof(nlp) <: ADNLPModel
+    set_adbackend!(nlp, ReverseDiffAD(nlp.meta.nvar, nlp.f))
+  elseif typeof(nlp) <: ADNLSModel
+    function F(x; nequ = nlp.nls_meta.nequ)
+      Fx = similar(x, nequ)
+      nlp.F!(Fx, x)
+      return Fx
+    end
+    set_adbackend!(nlp, ReverseDiffAD(nlp.meta.nvar, x -> sum(F(x) .^ 2)))
+  end
+  @test typeof(get_adbackend(nlp).gradient_backend) <: ADNLPModels.ReverseDiffADGradient
+  @test typeof(get_adbackend(nlp).hprod_backend) <: ADNLPModels.ReverseDiffADHvprod
+  @test typeof(get_adbackend(nlp).hessian_backend) <: ADNLPModels.ReverseDiffADHessian
+  set_adbackend!(
+    nlp,
+    gradient_backend = ADNLPModels.ForwardDiffADGradient,
+    jtprod_backend = ADNLPModels.GenericForwardDiffADJtprod(),
+  )
+  @test typeof(get_adbackend(nlp).gradient_backend) <: ADNLPModels.ForwardDiffADGradient
+  @test typeof(get_adbackend(nlp).hprod_backend) <: ADNLPModels.ReverseDiffADHvprod
+  @test typeof(get_adbackend(nlp).jtprod_backend) <: ADNLPModels.GenericForwardDiffADJtprod
+  @test typeof(get_adbackend(nlp).hessian_backend) <: ADNLPModels.ReverseDiffADHessian
+end
